+++ domain/src/main/java/tachiyomi/domain/chapter/model/Chapter.kt
@@ -1,5 +1,7 @@
 package tachiyomi.domain.chapter.model
 
+import kotlinx.serialization.json.JsonObject
+
 data class Chapter(
     val id: Long,
     val mangaId: Long,
@@ -15,6 +17,7 @@ data class Chapter(
     val scanlator: String?,
     val lastModifiedAt: Long,
     val version: Long,
+    val memo: JsonObject = JsonObject(emptyMap()),
 ) {
     val isRecognizedNumber: Boolean
         get() = chapterNumber >= 0f
@@ -26,6 +29,7 @@ data class Chapter(
             dateUpload = other.dateUpload,
             chapterNumber = other.chapterNumber,
             scanlator = other.scanlator?.ifBlank { null },
+            memo = other.memo,
         )
     }
 
@@ -45,6 +49,7 @@ data class Chapter(
             scanlator = null,
             lastModifiedAt = 0,
             version = 1,
+            memo = JsonObject(emptyMap()),
         )
     }
 }